requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
email-validator
python-dotenv
requests
ultralytics
easyocr
torch
torchvision
torchaudio
scikit-learn
python-multipart
httpx
geopy